+++ back/src/main/java/com/example/gestion_des_epi/gestion_epi/repository/PosteEpiRepository.java
@@ -1,7 +1,9 @@
 package com.example.gestion_des_epi.gestion_epi.repository;
 
-import com.example.gestion_des_epi.gestion_epi.model.PosteEpi;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PosteEpiRepository  extends JpaRepository<PosteEpi, Long> {
+import com.example.gestion_des_epi.gestion_epi.model.PosteEpi;
+
+
+public interface PosteEpiRepository  extends JpaRepository<PosteEpi, Integer> {
 }
